requirements.txt
google-generativeai==0.8.3
streamlit==1.41.1
python-dotenv